requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Werkzeug==3.0.1
email-validator==2.1.0
reportlab==4.1.0
itsdangerous==2.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
